requirements.txt
tensorflow-gpu
tensorflow_hub
numpy
Pillow
slackbot
slackclient
picamera